sem_2/lab_12/requirements.txt
tkinter
openpyxl
python-docx
